fix precision

+++ rest/result.go
@@ -184,7 +184,7 @@ type PriceFilter struct {
 }
 
 type LotSizeFilter struct {
-	MaxTradingQty int     `json:"max_trading_qty"`
+	MaxTradingQty float64 `json:"max_trading_qty"`
 	MinTradingQty float64 `json:"min_trading_qty"`
 	QtyStep       float64 `json:"qty_step"`
 }
